Add authentication, authorization and fix old specs

+++ app/controllers/api/bookings_controller.rb
@@ -3,17 +3,18 @@ module Api
     before_action :set_booking, only: [:show, :update, :destroy]
 
     def index
-      # render json: BookingSerializer.render(Booking.all, root: :bookings), status: :ok
-      render json: serialize(Booking.all, root: :bookings)
+      bookings = policy_scope(Booking)
+      render json: serialize(bookings, root: :bookings)
     end
 
     def show
-      # render json: BookingSerializer.render(booking, root: :booking), status: :ok
+      authorize booking
       render json: serialize(booking, root: :booking)
     end
 
     def create
-      new_booking = Booking.new(booking_params)
+      new_booking = current_user.bookings.build(booking_params)
+      authorize new_booking
       if new_booking.save
         render json: BookingSerializer.render(new_booking, root: :booking), status: :created
       else
@@ -22,6 +23,7 @@ module Api
     end
 
     def update
+      authorize booking
       if booking.update(booking_params)
         render json: BookingSerializer.render(booking, root: :booking), status: :ok
       else
@@ -30,6 +32,7 @@ module Api
     end
 
     def destroy
+      authorize booking
       booking.destroy
       head :no_content
     end
@@ -37,15 +40,18 @@ module Api
     private
 
     def set_booking
-      render_not_found unless booking
+      @booking = Booking.find_by(id: params[:id])
+      render_not_found unless @booking
     end
 
-    def booking
-      @booking ||= Booking.find_by(id: params[:id])
-    end
+    attr_reader :booking
 
     def booking_params
-      params.require(:booking).permit(:no_of_seats, :seat_price, :user_id, :flight_id)
+      if current_user.admin?
+        params.require(:booking).permit(:no_of_seats, :seat_price, :flight_id, :user_id)
+      else
+        params.require(:booking).permit(:no_of_seats, :seat_price, :flight_id)
+      end
     end
   end
 end
