+++ app/controllers/api/companies_controller.rb
@@ -1,18 +1,18 @@
 module Api
   class CompaniesController < ApplicationController
     before_action :set_company, only: [:show, :update, :destroy]
+    skip_before_action :authenticate_user, only: [:index, :show]
 
     def index
-      # render json: CompanySerializer.render(Company.all, root: :companies), status: :ok
       render json: serialize(Company.all, root: :companies)
     end
 
     def show
-      # render json: CompanySerializer.render(company, root: :company), status: :ok
       render json: serialize(company, root: :company)
     end
 
     def create
+      authorize Company
       new_company = Company.new(company_params)
       if new_company.save
         render json: CompanySerializer.render(new_company, root: :company), status: :created
@@ -22,6 +22,7 @@ module Api
     end
 
     def update
+      authorize company
       if company.update(company_params)
         render json: CompanySerializer.render(company, root: :company), status: :ok
       else
@@ -30,6 +31,7 @@ module Api
     end
 
     def destroy
+      authorize company
       company.destroy
       head :no_content
     end
